requirements.txt
python-telegram-bot==21.6
groq>=0.9.0
apscheduler==3.10.4
pytz==2024.1